Ultimos ajustes

+++ Server/Server/Models/Cliente.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Server.Models
 {
@@ -14,7 +11,7 @@ namespace Server.Models
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Display(Name = "Id")]
         [Column("Id")]
-        public Int64 Id { get; set; }
+        public int Id { get; set; }
 
         [Display(Name = "Nome")]
         [Column("Nome")]
